+++ src/graphql_clj/validator.clj
@@ -15,6 +15,10 @@
             [graphql-clj.validator.rules.unique-fragment-names]
             [graphql-clj.validator.rules.unique-argument-names]
             [graphql-clj.validator.rules.provided-non-null-arguments]
+            [graphql-clj.validator.rules.no-unused-variables]
+            [graphql-clj.validator.rules.no-unused-fragments]
+            [graphql-clj.validator.rules.known-directives]
+            [graphql-clj.validator.rules.lone-anonymous-operation]
             [graphql-clj.visitor :as visitor]
             [graphql-clj.spec :as spec]
             [instaparse.core :as insta]
@@ -27,7 +31,8 @@
             graphql-clj.validator.rules.unique-argument-names/schema-rules]))
 
 (def second-pass-rules-statement
-  (flatten [graphql-clj.validator.rules.known-type-names/rules
+  (flatten [graphql-clj.validator.rules.lone-anonymous-operation/rules
+            graphql-clj.validator.rules.known-type-names/rules
             graphql-clj.validator.rules.known-argument-names/rules
             graphql-clj.validator.rules.known-fragment-names/rules
             graphql-clj.validator.rules.no-undefined-variables/rules
@@ -42,7 +47,10 @@
             graphql-clj.validator.rules.unique-operation-names/rules
             graphql-clj.validator.rules.unique-fragment-names/rules
             graphql-clj.validator.rules.unique-argument-names/statement-rules
-            graphql-clj.validator.rules.provided-non-null-arguments/rules]))
+            graphql-clj.validator.rules.provided-non-null-arguments/rules
+            graphql-clj.validator.rules.no-unused-variables/rules
+            graphql-clj.validator.rules.no-unused-fragments/rules
+            graphql-clj.validator.rules.known-directives/rules]))
 
 (defn- validate [visit-fn]
   (try (visit-fn)