+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file. This change
 - Make validation robust to order (can now reference types declared later in a statement or schema)
 - Validate unique variable, operation, input type, fragment, and argument names
 - Validate provided non-null arguments
+- Validate unused variables and fragments
+- Validate known directives
+- Validate lone anonymous operation
 
 ## [0.1.17] - 2016-10-14
 - Implement preliminary version of FieldsOnCorrectType validation